[RUMF-1337] report fetch on payload callback
